Fix cmd.exe
It seems that there is a race condition between the main loop of Qt, and all the setup I do. The only way to overcome this, is to setup the PTY when the window is properly up

This is done delaying the setup of the PTY on the next event poll.

Seems stable enough. I even managed to get git/bash running. But its very slow.